* commit 'da2b97b03b40c3895d3060ca5bc6a336537518cc':
  Improve handling of resendMode variable

+++ drvFGPDBApp/test/drvFGPDBTests.cpp
@@ -77,7 +77,7 @@ public:
                 disconnect(_)).WillOnce(Return(asynSuccess));
     testDrv = make_unique<drvFGPDB>(drvName, syncIO, UDPPortName,
                                     startupDiagFlags,
-                                    static_cast<uint32_t>(ResendMode::AfterCtlrRestart));
+                                    ResendMode::AfterCtlrRestart);
 
     if (udpPortStat)
       cout << drvName << " unable to create asyn UDP port: " << UDPPortName
